Añadir columna `idActuacionExterna` en la entidad `Ocupacion`.

+++ src/main/java/es/musicalia/gestmusica/ocupacion/Ocupacion.java
@@ -110,7 +110,7 @@ public class Ocupacion {
     private boolean soloMatinal;
 
     @Column(name = "id_actuacion_externa")
-    private Integer idActuacionExterna;
+    private Long idActuacionExterna;
 
 
 }
